blood-donation-ai/requirements.txt
crewai
crewai_tools
python-dotenv
langchain-google-genai
geopy